Get Django tests running
This commit contains a massive amount of changes that should be enough
to get tests running. These changes need to be cleaned up and
compartmentalized, along with removing a bunch of debug output left in.

Since Django tests rely on creating test "databases", we need to map
that concept to django-ibmi as well. IBM i only vaguely has the concept
of separate databases with both remote databases and iASPs, but neither
of those map to something that can be created ad-hoc. Django just needs
a container to keep its stuff in separate from the other users of the
DBMS, so we can map this nicely to a schema.

Add a document for setting up and running tests as well.