Implement datetime_cast_date_sql and datetime_cast_time_sql
These don't support timezone behavior yet. Not sure how, since the other
backend databases seem to have timezone functions built in. We could use
UTC offset like other code does, but that won't be correct based on the
DST settings in effect for the timestamp stored in the database (ie. we
can get the current UTC offset, but the UTC offset at the time stored in
the timestamp may be different).

Also remove datetime_cast_sql, which is unused since Django 1.9.